refactor(server): separate files
