+++ server/src/env.ts
@@ -1,0 +1,13 @@
+import dotenv from "dotenv";
+import z from "zod";
+
+dotenv.config();
+
+const env = z
+	.object({
+		PORT: z.string().default("3001"),
+		GITHUB_TOKEN: z.string(),
+	})
+	.parse(process.env);
+
+export default env;
